Made reset value be u128 to make every known architecture (and then some) fit

+++ racr_rust/src/register.rs
@@ -11,7 +11,7 @@ pub struct RegisterDefinition {
     pub description: Option<String>,
 
     pub size: usize,
-    pub reset_value: Option<usize>,
+    pub reset_value: Option<u128>,
 
     pub overlapping: bool,
